+++ R/zzz.R
@@ -1,29 +1,6 @@
 .onLoad <- function(libname, pkgname) {
-  # if data pkg is installed, load it first
-  if (
-    !is_on_cran() &&
-    length(find.package("NBDCtoolsData", quiet = TRUE)) > 0
-  ) {
-    utils::data(
-      list = c(
-        "lst_dds",
-        "lst_levels",
-        "lst_sessions"
-      ),
-      package = "NBDCtoolsData",
-      envir = asNamespace(pkgname)
-    )
-  }
-  if (is_on_cran()) {
-    purrr::walk(
-      c("lst_dds", "lst_levels", "lst_sessions"),
-      ~ readRDS(system.file(
-        "extdata", "meta_internal", glue::glue("{.x}.rds"),
-        package = pkgname
-      )) |>
-        assign(.x, value = _, envir = asNamespace(pkgname))
-    )
-  }
+  # create empty environment for caching metadata objects
+  options("NBDCtoolsData.env" = new.env())
 }
 
 .onAttach <- function(libname, pkgname) {
@@ -43,16 +20,5 @@
     #   "L Zhang, xxx & J Linkersdörfer. NBDCtools: xxx. 2025. xxx",
     # ))
     options(nbdctools_start_msg_displayed = TRUE)
-  }
-
-  if (!is_on_cran() && length(find.package("NBDCtoolsData", quiet = TRUE)) < 1) {
-    cli::cli_warn(
-      c(
-        "x" = glue::glue(
-          "The `NBDCtoolsData` package cannot be found. Please install it ",
-          "using: remotes::install_github('nbdc-datahub/NBDCtoolsData')"
-        )
-      )
-    )
   }
 }